+++ Sources/SuperwallKit/Paywall/View Controller/Web View/SWWebView.swift
@@ -65,9 +65,10 @@ class SWWebView: WKWebView {
     }
 
     let preferences = WKPreferences()
+
     if #available(iOS 15.0, *),
       !isMac {
-      preferences.isTextInteractionEnabled = false // ignore-xcode-12
+      preferences.isTextInteractionEnabled = featureFlags?.enableTextInteraction == true // ignore-xcode-12
     }
     preferences.javaScriptCanOpenWindowsAutomatically = true